+++ tests/test_cli.py
@@ -119,6 +119,22 @@ def netrc_from_environ(no_netrc, credentials):
 
 @pytest.mark.vcr
 @pytest.mark.scihub
+def test_cli_gnss(run_cli):
+    run_cli(
+        "--gnss",
+        "-s",
+        "20210201",
+        "-e",
+        "20210202",
+        "--producttype",
+        "AUX_POEORB",
+        "--query",
+        "platformserialidentifier=1B",
+    )
+
+
+@pytest.mark.vcr
+@pytest.mark.scihub
 def test_cli_geometry_alternatives(run_cli, geojson_string, wkt_string):
     run_cli("--geometry", geojson_string, "--end", "20200101", "--limit", "1")
     run_cli("--geometry", wkt_string, "--end", "20200101", "--limit", "1")
@@ -423,32 +439,6 @@ def test_repeated_keywords(run_cli):
 @pytest.mark.scihub
 def test_name_search_empty(run_cli):
     run_cli("--name", "", must_raise=ValueError)
-
-
-@pytest.mark.vcr
-@pytest.mark.scihub
-def test_option_hierarchy(run_cli, geojson_path):
-    # expected hierarchy is producttype > instrument > platform from most to least specific
-    result = run_cli(
-        "--geometry",
-        geojson_path,
-        "--url",
-        "https://apihub.copernicus.eu/apihub/",
-        "-s",
-        "20161201",
-        "-e",
-        "20161202",
-        "--sentinel",
-        "1",
-        "--instrument",
-        "SAR-C SAR",
-        "--producttype",
-        "S2MSI1C",
-    )
-
-    # Check that all returned products are of type 'S2MSI1C'
-    assert len(result.products) > 0
-    assert all("Instrument: MSI, Satellite: Sentinel-2" in p for p in result.products)
 
 
 @pytest.mark.vcr